@@ -1,22 +1,22 @@
 {
     "name": "vanio/easy-mailer",
-    "description": "Easy peasy e-mail sending from PHP applications.",
+    "description": "Easy e-mail sending from PHP applications.",
     "keywords": ["e-mail", "e-mail sending"],
     "homepage": "https://github.com/vaniocz/easy-mailer",
     "license": "MIT",
     "authors": [
-        {"name": "Adam Hojka"}
+        {"name": "Vanio"}
     ],
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.1",
         "pelago/emogrifier": "^1.2 || ^2.0 || ^3.0 || ^4.0 || ^5.0",
         "html2text/html2text": "^4.0",
         "swiftmailer/swiftmailer": "^5.4 || ^6.0",
-        "twig/twig": "^1.24 || ^2.0 || ^3.0"
+        "twig/twig": "^2.7 || ^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.5",
-        "vanio/coding-standards": "^0.1"
+        "phpunit/phpunit": "^9.5",
+        "vanio/coding-standards": "^0.4"
     },
     "autoload": {
         "psr-4": {"Vanio\\EasyMailer\\": "src/"},
